♻️ Refactor: EstimateCard 수정
- 날짜 계산 함수 utils파일로 빼놓음
- 희망날짜 오류 수정
